Update Available Options in DateTime Class
Update the DateTime class to allow for three different inclusivity modes: ignored, included and required.

If a rule matches the start date, IGNORED will prevent the start date from being added to the list of occurrences; if the rule matches but INCLUDED is set, the start date will be added to the list of occurrences; however if the rule does not match the start date and REQUIRED is set, an InvalidStartDate exception will be thrown.